move each command's entry point logic to their own files
